+++ src/main/java/tk/cth451/transitrailmod/EnumGuiTypes.java
@@ -1,0 +1,5 @@
+package tk.cth451.transitrailmod;
+
+public enum EnumGuiTypes {
+	TICKET_MACHINE;
+}