+++ src/main/java/tk/cth451/transitrailmod/TransitRailMod.java
@@ -2,11 +2,14 @@ package tk.cth451.transitrailmod;
 
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
+import net.minecraftforge.fml.common.Mod.Instance;
 import net.minecraftforge.fml.common.SidedProxy;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+import net.minecraftforge.fml.common.network.NetworkRegistry;
 import tk.cth451.transitrailmod.init.ModBlocks;
+import tk.cth451.transitrailmod.init.ModGuiHandler;
 import tk.cth451.transitrailmod.init.ModItems;
 import tk.cth451.transitrailmod.init.ModRecipe;
 import tk.cth451.transitrailmod.init.ModTileEntities;
@@ -15,6 +18,9 @@ import tk.cth451.transitrailmod.proxy.CommonProxy;
 @Mod(modid = References.MOD_ID, version = References.VERSION, name = References.VERSION)
 public class TransitRailMod
 {
+	@Instance(References.MOD_ID)
+	public static TransitRailMod instance;
+	
 	@SidedProxy(clientSide = References.CLIENT_PROXY_CLASS, serverSide = References.SERVER_PROXY_CLASS)
 	public static CommonProxy proxy;
 	
@@ -36,6 +42,7 @@ public class TransitRailMod
 		ModRecipe.addItemsRecipe();
 		ModRecipe.addBlocksRecipe();
 		proxy.registerRenders();
+		NetworkRegistry.INSTANCE.registerGuiHandler(TransitRailMod.instance, new ModGuiHandler());
 	}
 	
 	@EventHandler